feat: add edit category

+++ categories/CategoriesController.js
@@ -4,7 +4,6 @@ const router = express.Router();
 const Category = require('./Category');
 
 const slugify = require('slugify');
-const {route} = require('express/lib/application');
 
 router.get('/admin/categories', async (request, response) => {
   const categories = await Category.findAll({raw: true});
@@ -18,7 +17,7 @@ router.get('/admin/categories/create', (request, response) => {
   return response.render('admin/categories/create');
 });
 
-router.post('/categories/save', async (request, response) => {
+router.post('/admin/categories/save', async (request, response) => {
   const {title} = request.body;
 
   if (!title) return response.redirect('/admin/categories/create');
@@ -31,10 +30,44 @@ router.post('/categories/save', async (request, response) => {
   return response.redirect('/admin/categories');
 });
 
-router.get('/categories/delete/:id', async (request, response) => {
+router.get('/admin/categories/edit/:id', async (request, response) => {
   const {id} = request.params;
 
-  if (!id) return response.redirect('/admin/categories');
+  if (!id || isNaN(id)) return response.redirect('/admin/categories');
+
+  const category = await Category.findByPk(id);
+
+  if (!category) return response.redirect('/admin/categories');
+
+  return response.render('admin/categories/edit', {
+    category,
+  });
+});
+
+router.post('/admin/categories/update', async (request, response) => {
+  const {id, title} = request.body;
+
+  if (!id || !title) return response.redirect('/admin/categories');
+
+  await Category.update(
+    {
+      title,
+      slug: slugify(title),
+    },
+    {
+      where: {
+        id,
+      },
+    },
+  );
+
+  return response.redirect('/admin/categories');
+});
+
+router.post('/admin/categories/delete', async (request, response) => {
+  const {id} = request.body;
+
+  if (!id || isNaN(id)) return response.redirect('/admin/categories');
 
   await Category.destroy({
     where: {
